Fix: Rename feedback types
