+++ apps/web/components/user-feedback/feedback-badge.tsx
@@ -1,5 +1,5 @@
-import { Sentiment } from "@/shared/sentiment-types";
 import { Badge } from "@repo/ui/components/badge";
+import { type Sentiment } from "@/types/sentimentAnalysisResult";
 
 const sentimentColors: Record<Sentiment, string> = {
   positive: "bg-green-200 text-green-800",
