--- apps/web/shared/sentiment-types.ts
@@ -1,1 +1,0 @@
-export type Sentiment = "positive" | "neutral" | "negative" | "unknown";
