+++ apps/web/types/sentimentAnalysisResult.ts
@@ -1,0 +1,16 @@
+export type Sentiment = "positive" | "neutral" | "negative" | "unknown";
+
+export type SentimentAnalysisResult = {
+  id: string;
+  content: string;
+  sentiment: Sentiment;
+  confidence: number;
+  summary: string;
+  created_at: string;
+};
+
+export type SentimentAnalysisGroup = {
+  id: string;
+  groupName: string;
+  items: SentimentAnalysisResult[];
+};
